chore(weave): extract fields.py from calls_query_builder
Extract the field class hierarchy, field registry, summary handlers,
and related constants into a new fields.py module. This is the first
step in breaking up the 3008-line calls_query_builder.py monolith.

Extracted components:
- Field classes: QueryBuilderField, CallsMergedField, CallsMergedAggField,
  CallsMergedDynamicField, CallsMergedSummaryField, CallsMergedFeedbackPayloadField,
  CallsMergedQueueItemField, AggregatedDataSizeField, QueryBuilderDynamicField
- Field registry: ALLOWED_CALL_FIELDS, get_field_by_name()
- Summary handlers: status, latency_ms, trace_name
- Constants: DATETIME_COLUMN_FIELDS, DISALLOWED_FILTERING_FIELDS, etc.
